feat: Add script to create a new user with sudo privileges
This commit adds a new script, 1-createUser.sh, which allows the creation of a new user with sudo privileges. The script checks if it is running as root or with sudo, prompts for the new user's name and password, creates the user, sets the password, and grants sudo privileges. This script will be useful for setting up new user accounts on the system.